Minor Linux fix
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@4548 42547f41-83fc-0310-9aef-9668e019ab9e

+++ modules/xglab_dante/i_dante.cpp
@@ -200,7 +200,7 @@ mxi_dante_open( MX_RECORD *record )
 	MX_DANTE *dante = NULL;
 	long dimension[3];
 	size_t dimension_sizeof[3];
-	int i, j;
+	unsigned long i, j;
 	unsigned long max_chain_boards;
 	unsigned long flags;
